backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
PyMuPDF==1.24.3
pydantic==2.7.0
python-dotenv==1.0.1
llama-cpp-python>=0.3.4
